refactor(hub-common): change helper file name
affects: @esri/hub-common
